init Payment with projectId and rawPaymentMethods
